--- data/src/main/java/com/example/data/model/signin/request/SignInRequest.kt
@@ -1,5 +1,0 @@
-package com.example.data.model.signin.request
-
-data class SignInRequest(
-    val code: String
-)
