--- data/src/main/java/com/example/data/source/SignInDataSource.kt
@@ -1,10 +1,0 @@
-package com.example.data.source
-
-import com.example.data.model.signin.response.SignInResponse
-import kotlinx.coroutines.flow.Flow
-
-interface SignInDataSource {
-    suspend fun signIn(
-        code: String
-    ): Flow<SignInResponse>
-}
