+++ data/src/main/java/com/goms/data/model/signin/request/SignInRequest.kt
@@ -1,0 +1,5 @@
+package com.goms.data.model.signin.request
+
+data class SignInRequest(
+    val code: String
+)
