--- domain/src/main/java/com/example/domain/data/signin/request/SignInRequestData.kt
@@ -1,5 +1,0 @@
-package com.example.domain.data.signin.request
-
-data class SignInRequestData(
-    val code: String
-)
