--- domain/src/main/java/com/example/domain/data/signin/response/SignInResponseData.kt
@@ -1,10 +1,0 @@
-package com.example.domain.data.signin.response
-
-import java.time.LocalDateTime
-
-data class SignInResponseData(
-    val accessToken: String,
-    val refreshToken: String,
-    val accessTokenExpiredAt: LocalDateTime,
-    val refreshTokenExpiredAt: LocalDateTime
-)
