--- domain/src/main/java/com/example/domain/repository/SignInRepository.kt
@@ -1,10 +1,0 @@
-package com.example.domain.repository
-
-import com.example.domain.data.signin.response.SignInResponseData
-import kotlinx.coroutines.flow.Flow
-
-interface SignInRepository {
-    suspend fun signIn(
-        code: String
-    ): Flow<SignInResponseData>
-}
