+++ domain/src/main/java/com/goms/domain/data/signin/request/SignInRequestData.kt
@@ -1,0 +1,5 @@
+package com.goms.domain.data.signin.request
+
+data class SignInRequestData(
+    val code: String
+)
