--- presentation/src/main/java/com/example/presentation/di/GomsApplication.kt
@@ -1,7 +1,0 @@
-package com.example.presentation.di
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class GomsApplication: Application()
