--- presentation/src/main/java/com/example/presentation/di/module/DataSourceModule.kt
@@ -1,18 +1,0 @@
-package com.example.presentation.di.module
-
-import com.example.data.source.SignInDataSource
-import com.example.data.source.SignInDataSourceImpl
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module
-@InstallIn(SingletonComponent::class)
-abstract class DataSourceModule {
-
-    @Binds
-    abstract fun provideSignInDataSource(
-        signInDataSourceImpl: SignInDataSourceImpl
-    ): SignInDataSource
-}
