--- presentation/src/main/java/com/example/presentation/di/module/NetWorkModule.kt
@@ -1,60 +1,0 @@
-package com.example.presentation.di.module
-
-import com.example.data.api.SignInService
-import com.google.gson.GsonBuilder
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import java.util.concurrent.TimeUnit
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object NetWorkModule {
-    const val BASE_URL = "http://10.82.20.18:8080/"
-
-    @Provides
-    @Singleton
-    fun provideOkhttpClient(): OkHttpClient {
-        return OkHttpClient.Builder()
-            .addInterceptor(HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BODY))
-            .connectTimeout(30, TimeUnit.SECONDS)
-            .readTimeout(30, TimeUnit.SECONDS)
-            .writeTimeout(30, TimeUnit.SECONDS)
-            .build()
-    }
-
-    @Provides
-    @Singleton
-    fun provideGsonConverterFactory(): GsonConverterFactory {
-        return GsonConverterFactory.create(
-            GsonBuilder()
-                .setLenient()
-                .create()
-        )
-    }
-
-    @Provides
-    @Singleton
-    fun provideRetrofitInstance(
-        okHttpClient: OkHttpClient,
-        gsonConverterFactory: GsonConverterFactory
-    ): Retrofit {
-        return Retrofit.Builder()
-            .baseUrl(BASE_URL)
-            .client(okHttpClient)
-            .addConverterFactory(gsonConverterFactory)
-            .build()
-    }
-
-    @Provides
-    @Singleton
-    fun provideSignInService(retrofit: Retrofit): SignInService {
-        return retrofit.create(SignInService::class.java)
-    }
-}
