+++ presentation/src/main/java/com/goms/presentation/di/GomsApplication.kt
@@ -1,0 +1,7 @@
+package com.goms.presentation.di
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class GomsApplication: Application()
