+++ presentation/src/main/java/com/goms/presentation/di/module/DataSourceModule.kt
@@ -1,0 +1,18 @@
+package com.goms.presentation.di.module
+
+import com.goms.data.source.SignInDataSource
+import com.goms.data.source.SignInDataSourceImpl
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+abstract class DataSourceModule {
+
+    @Binds
+    abstract fun provideSignInDataSource(
+        signInDataSourceImpl: SignInDataSourceImpl
+    ): SignInDataSource
+}
